Fixed disrepancies between a few getter and setters.

+++ modules/entity_spell_system/data/loot/loot_data_base.cpp
@@ -27,7 +27,7 @@ SOFTWARE.
 int LootDataBase::get_loot_db_size() const {
 	return _loot_dbs.size();
 }
-void LootDataBase::set_loot_db_size(const float value) {
+void LootDataBase::set_loot_db_size(const int value) {
 	_loot_dbs.resize(value);
 }
